requirements.txt
azure-functions==1.7.2
certifi==2021.10.8
charset-normalizer==2.0.7
Flask==2.0.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
requests==2.26.0
urllib3==1.26.7
pandas
xlrd